refactor(All) : domain 카테고리 정리(package)
