+++ src/main/java/laddergameplay/LadderGameMain.java
@@ -1,6 +1,10 @@
 package laddergameplay;
 
-import laddergameplay.domain.*;
+import laddergameplay.domain.datas.People;
+import laddergameplay.domain.datas.Results;
+import laddergameplay.domain.dto.WinningResult;
+import laddergameplay.domain.ladder.Ladder;
+import laddergameplay.domain.ladder.LadderGame;
 import laddergameplay.exception.CustomException;
 import laddergameplay.strategy.LineStrategy;
 import laddergameplay.strategy.RandomLineStrategy;
