+++ src/main/java/laddergameplay/view/ResultView.java
@@ -1,6 +1,12 @@
 package laddergameplay.view;
 
-import laddergameplay.domain.*;
+import laddergameplay.domain.data.Person;
+import laddergameplay.domain.data.Result;
+import laddergameplay.domain.datas.People;
+import laddergameplay.domain.datas.Results;
+import laddergameplay.domain.dto.WinningResult;
+import laddergameplay.domain.ladder.Ladder;
+import laddergameplay.domain.ladder.Line;
 
 import java.util.Map;
